refactor: Car, MoveStrategy, Setting, VictoryCondition 클래스 domain 패키지 안으로 이동
