+++ src/main/kotlin/racingcar/domain/Setting.kt
@@ -1,3 +1,3 @@
-package racingcar
+package racingcar.domain
 
 data class Setting(val nameOfCars: List<String>, val numberOfLab: Int = 5)
