+++ src/main/kotlin/racingcar/domain/VictoryCondition.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 enum class VictoryCondition {
     Longest {
