+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,6 +1,6 @@
 package racingcar.view
 
-import racingcar.Car
+import racingcar.domain.Car
 
 class ResultView(private val progressCharacter: String = "-") {
