+++ src/test/kotlin/racingcar/RacingGameTest.kt
@@ -2,6 +2,7 @@ package racingcar
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
+import racingcar.domain.Setting
 
 class RacingGameTest : BehaviorSpec({
