+++ src/test/kotlin/racingcar/domain/MoveStrategyTest.kt
@@ -1,9 +1,10 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.domain.MoveStrategy
 
 internal class MoveStrategyTest {
