+++ src/test/kotlin/racingcar/view/InputViewTest.kt
@@ -3,7 +3,7 @@ package racingcar.view
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import racingcar.Setting
+import racingcar.domain.Setting
 
 class InputViewTest : BehaviorSpec({
